refactor(tests): Move test util functions into conftest
